Switch new MIPRO to MIPROv2
Also switch all the old test cases back
