+++ dspy/teleprompt/__init__.py
@@ -4,6 +4,7 @@ from .ensemble import *
 from .finetune import *
 from .knn_fewshot import *
 from .mipro_optimizer import MIPRO
+from .mipro_optimizer_v2 import MIPROv2
 from .random_search import *
 from .signature_opt import SignatureOptimizer
 from .signature_opt_bayesian import BayesianSignatureOptimizer
